+++ spec/osm/member_spec.rb
@@ -41,7 +41,6 @@ describe "Member" do
       :custom9 => 'Custom Field 9',
       :grouping_id => '3',
       :grouping_leader => 0,
-      :grouping_label => 'Grouping',
       :joined => '2006-01-07',
       :age => '06/07',
       :joined_years => 1,
@@ -83,7 +82,6 @@ describe "Member" do
     member.custom8.should == 'Custom Field 8'
     member.custom9.should == 'Custom Field 9'
     member.grouping_id.should == 3
-    member.grouping_label.should == 'Grouping'
     member.grouping_leader.should == 0
     member.joined.should == Date.new(2006, 1, 7)
     member.age.should == '06/07'
@@ -274,7 +272,7 @@ describe "Member" do
       }
 
       Osm::Term.stub(:get_for_section) { [] }
-      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":"ok","scoutid":1}'}) }
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":"ok","scoutid":1}'}) }
       member.create(@api).should be_true
       member.id.should == 1
     end
@@ -291,49 +289,48 @@ describe "Member" do
         :grouping_leader => 0,
       )
 
-      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":"ok","scoutid":-1}'}) }
+      HTTParty.should_receive(:post) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":"ok","scoutid":-1}'}) }
       member.create(@api).should be_false
     end
 
 
     it "Update in OSM (succeded)" do
-      member = Osm::Member.new(
-        :id => 1,
-        :section_id => 2,
-        :first_name => 'First',
-        :last_name => 'Last',
-        :email1 => 'email1@example.com',
-        :email2 => 'email2@example.com',
-        :email3 => 'email3@example.com',
-        :email4 => 'email4@example.com',
-        :phone1 => '11111 111111',
-        :phone2 => '222222',
-        :phone3 => '+33 3333 333333',
-        :phone4 => '4444 444 444',
-        :address => '1 Some Road',
-        :address2 => 'Address 2',
-        :date_of_birth => '2000-01-02',
-        :started => '2006-01-02',
-        :joined => '2006-01-03',
-        :parents => 'John and Jane Doe',
-        :notes => 'None',
-        :medical => 'Nothing',
-        :religion => 'Unknown',
-        :school => 'Some School',
-        :ethnicity => 'Yes',
-        :subs => 'Upto end of 2007',
-        :custom1 => 'Custom Field 1',
-        :custom2 => 'Custom Field 2',
-        :custom3 => 'Custom Field 3',
-        :custom4 => 'Custom Field 4',
-        :custom5 => 'Custom Field 5',
-        :custom6 => 'Custom Field 6',
-        :custom7 => 'Custom Field 7',
-        :custom8 => 'Custom Field 8',
-        :custom9 => 'Custom Field 9',
-        :grouping_id => 3,
-        :grouping_leader => 0,
-      )
+      member = Osm::Member.new()
+      member.id = 1
+      member.section_id = 2
+      member.first_name = 'First'
+      member.last_name = 'Last'
+      member.email1 = 'email1@example.com'
+      member.email2 = 'email2@example.com'
+      member.email3 = 'email3@example.com'
+      member.email4 = 'email4@example.com'
+      member.phone1 = '11111 111111'
+      member.phone2 = '222222'
+      member.phone3 = '+33 3333 333333'
+      member.phone4 = '4444 444 444'
+      member.address = '1 Some Road'
+      member.address2 = 'Address 2'
+      member.date_of_birth = '2000-01-02'
+      member.started = '2006-01-02'
+      member.joined = '2006-01-03'
+      member.parents = 'John and Jane Doe'
+      member.notes = 'None'
+      member.medical = 'Nothing'
+      member.religion = 'Unknown'
+      member.school = 'Some School'
+      member.ethnicity = 'Yes'
+      member.subs = 'Upto end of 2007'
+      member.custom1 = 'Custom Field 1'
+      member.custom2 = 'Custom Field 2'
+      member.custom3 = 'Custom Field 3'
+      member.custom4 = 'Custom Field 4'
+      member.custom5 = 'Custom Field 5'
+      member.custom6 = 'Custom Field 6'
+      member.custom7 = 'Custom Field 7'
+      member.custom8 = 'Custom Field 8'
+      member.custom9 = 'Custom Field 9'
+      member.grouping_id = 3
+      member.grouping_leader = 0
 
       url = 'https://www.onlinescoutmanager.co.uk/users.php?action=updateMember&dateFormat=generic'
       body_data = {
@@ -384,7 +381,7 @@ describe "Member" do
             'column' => column,
             'value' => value,
             'sectionid' => member.section_id,
-          }}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>body}) }
+          }}) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>body}) }
         end
       end
       HTTParty.should_receive(:post).with('https://www.onlinescoutmanager.co.uk/users.php?action=updateMemberPatrol', {:body => {
@@ -396,7 +393,59 @@ describe "Member" do
         'patrolid' => member.grouping_id,
         'pl' => member.grouping_leader,
         'sectionid' => member.section_id,
-      }}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>body}) }
+      }}) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>body}) }
+      Osm::Term.stub(:get_for_section) { [] }
+
+      member.update(@api).should be_true
+    end
+
+    it "Update in OSM (only updated fields)" do
+      member = Osm::Member.new(
+        :id => 1,
+        :section_id => 2,
+        :date_of_birth => '2000-01-02',
+        :started => '2006-01-02',
+        :joined => '2006-01-03',
+        :grouping_leader => 0,
+      )
+      member.first_name = 'First'
+      member.last_name = 'Last'
+      member.grouping_id = 3
+
+      url = 'https://www.onlinescoutmanager.co.uk/users.php?action=updateMember&dateFormat=generic'
+      body_data = {
+        'firstname' => 'First',
+        'lastname' => 'Last',
+        'patrolid' => 3,
+        'patrolleader' => 0,
+      }
+      body = (body_data.inject({}) {|h,(k,v)| h[k]=v.to_s; h}).to_json
+
+      body_data.each do |column, value|
+        unless ['patrolid', 'patrolleader'].include?(column)
+          HTTParty.should_receive(:post).with(url, {:body => {
+            'apiid' => @CONFIGURATION[:api][:osm][:id],
+            'token' => @CONFIGURATION[:api][:osm][:token],
+            'userid' => 'user_id',
+            'secret' => 'secret',
+            'scoutid' => member.id,
+            'column' => column,
+            'value' => value,
+            'sectionid' => member.section_id,
+          }}) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>body}) }
+        end
+      end
+      HTTParty.should_receive(:post).with('https://www.onlinescoutmanager.co.uk/users.php?action=updateMemberPatrol', {:body => {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'scoutid' => member.id,
+        'patrolid' => member.grouping_id,
+        'pl' => member.grouping_leader,
+        'sectionid' => member.section_id,
+      }}) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>body}) }
+      Osm::Term.stub(:get_for_section) { [] }
 
       member.update(@api).should be_true
     end
@@ -405,7 +454,6 @@ describe "Member" do
       member = Osm::Member.new(
         :id => 1,
         :section_id => 2,
-        :first_name => 'First',
         :last_name => 'Last',
         :date_of_birth => '2000-01-02',
         :started => '2006-01-02',
@@ -413,9 +461,67 @@ describe "Member" do
         :grouping_id => '3',
         :grouping_leader => 0,
       )
+      member.first_name = 'First'
 
-      HTTParty.stub(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{}'}) }
+      HTTParty.stub(:post) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{}'}) }
       member.update(@api).should be_false
+    end
+
+    describe "Get My.SCOUT link" do
+      
+      before :each do
+        @member = Osm::Member.new(
+          :id => 1,
+          :section_id => 2,
+          :first_name => 'First',
+          :last_name => 'Last',
+          :date_of_birth => '2000-01-02',
+          :started => '2006-01-02',
+          :joined => '2006-01-03',
+          :grouping_id => '3',
+          :grouping_leader => 0,
+        )
+      end
+
+      it "Default" do
+        url = 'https://www.onlinescoutmanager.co.uk/api.php?action=getMyScoutKey&sectionid=2&scoutid=1'
+        HTTParty.should_receive(:post).with(url, {:body => {
+          'apiid' => @CONFIGURATION[:api][:osm][:id],
+          'token' => @CONFIGURATION[:api][:osm][:token],
+          'userid' => 'user_id',
+          'secret' => 'secret',
+        }}) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"ok":true,"key":"KEY-HERE"}'}) }
+        @member.myscout_link(@api).should == 'https://www.onlinescoutmanager.co.uk/parents/badges.php?sc=1&se=2&c=KEY-HERE'
+      end
+
+      it "Payments" do
+        url = 'https://www.onlinescoutmanager.co.uk/api.php?action=getMyScoutKey&sectionid=2&scoutid=1'
+        FakeWeb.register_uri(:post, url, :body => '{"ok":true,"key":"KEY-HERE"}')
+        @member.myscout_link(@api, :payments).should == 'https://www.onlinescoutmanager.co.uk/parents/payments.php?sc=1&se=2&c=KEY-HERE'
+      end
+
+      it "Events" do
+        url = 'https://www.onlinescoutmanager.co.uk/api.php?action=getMyScoutKey&sectionid=2&scoutid=1'
+        FakeWeb.register_uri(:post, url, :body => '{"ok":true,"key":"KEY-HERE"}')
+        @member.myscout_link(@api, :events).should == 'https://www.onlinescoutmanager.co.uk/parents/events.php?sc=1&se=2&c=KEY-HERE'
+      end
+
+      it "Programme" do
+        url = 'https://www.onlinescoutmanager.co.uk/api.php?action=getMyScoutKey&sectionid=2&scoutid=1'
+        FakeWeb.register_uri(:post, url, :body => '{"ok":true,"key":"KEY-HERE"}')
+        @member.myscout_link(@api, :programme).should == 'https://www.onlinescoutmanager.co.uk/parents/programme.php?sc=1&se=2&c=KEY-HERE'
+      end
+
+      it "Badges" do
+        url = 'https://www.onlinescoutmanager.co.uk/api.php?action=getMyScoutKey&sectionid=2&scoutid=1'
+        FakeWeb.register_uri(:post, url, :body => '{"ok":true,"key":"KEY-HERE"}')
+        @member.myscout_link(@api, :badges).should == 'https://www.onlinescoutmanager.co.uk/parents/badges.php?sc=1&se=2&c=KEY-HERE'
+      end
+
+      it "Invalid" do
+        expect{ @member.myscout_link(@api, :invalid) }.to raise_error(Osm::ArgumentIsInvalid)
+      end
+
     end
 
   end